File checking

+++ src/steganography/extract_cmd.py
@@ -2,19 +2,22 @@ from __future__ import annotations
 
 from argparse import Namespace
 
-from PIL import Image
+from PIL import Image, UnidentifiedImageError
 
 from steganography.decode import decode_file_from_image
 
 
 def extract(args: Namespace) -> int:
     image_path = args.file
-    # Check if it is a file
+    if not image_path.exists():
+        raise FileNotFoundError(f"The image {image_path} does not exists")
     encryption_key = args.password
     hashing = not args.no_hash
-    image = Image.open(image_path)
+    try:
+        image = Image.open(image_path)
+    except UnidentifiedImageError:
+        raise FileNotFoundError(f"{image_path} is not a valid Image")
 
-    # Check for Imagetype and ImageMode errors
     # Check for false user input ( ho_hash, password, etc. )
     file_name, file_bytes = decode_file_from_image(image, encryption_key, hashing)
     with open(file_name, "wb") as f:
